ml/requirements.txt
# ML/RL Requirements for Pose2Play Rehabilitation Model

# Core ML libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep Learning
torch>=1.10.0
tensorflow>=2.8.0

# Reinforcement Learning
gymnasium>=0.29.0
stable-baselines3>=2.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Data processing
scipy>=1.7.0

# Model export (optional)
onnx>=1.12.0
onnxruntime>=1.12.0
tensorflowjs>=3.18.0

# Progress bars
tqdm>=4.62.0

# API Server
flask>=2.3.0
flask-cors>=4.0.0

# Jupyter (for experimentation)
jupyter>=1.0.0
ipython>=7.30.0
